UI improvements, flash-when-ready button, bugfixes, etc. (#16)
* Show error state for all errors

* Show error state for all errors

* Move success check to correct location

* Check for blank output when resetting

as opposed to checking for one specific error

* Check for correct output when erasing

* Homogenize success checks

* Create flash_when_ready variable

* Remove code to handle the nonexistant reset button

* Removed code to handle old nonexistant close button

* Cache the JQuery selectors

Better for performance and (IMO) readability

* Update Bootstrap

* Offload  flash button styling to Bootstrap

The button will still be green when disabled this way. It always bugged
me how the button would be red after a successful flash. It made it look
like something was wrong.

* Disable the menubar when running from source

* Fix enableFlashButton method

* Added a button to flash when ready

Closes #3

* Get bootstrap from npm

* Fix issue with flash button

During the flash when ready state, the flash button would become enabled
once the keyboard was detected.

* Hide fwr button once the board is detected

* Add electron-debug dependency

Only active when in dev mode. Will allow the use of keyboard shortcuts
to open the dev tools and such.
https://www.npmjs.com/package/electron-debug

* Enable electron-debug

* Fix for Windows packaged version

* Correct dfu_location synchronously

This code wasn't working because it wouldn't change dfu_location until
after dfu_location was accessed.

* Add script to build a Windows package on Windows

* Fixed bad dfu path on Windows packaged version

* Improve FWR button alignment

* Improve window sizing on Windows

tested on Windows 7 and 10

* Improve error on 32-bit Windows

* Add more detail about driver installation to log

* Revert "Improve error on 32-bit Windows"

This reverts commit c66373a69003561c85777afe9401d620990f4b42.

* Run npm install before packaging using batch file

Get all updates

* Change disabled button styling from green to gray

* If the keyboard is unplugged, reenable FWR button

I also attempted to make some progress towards better supporting typing
the hex path manually, but that still needs work.